Move Dockerfile back to top-level dir
Can't get automated build to work with it under docker/, as the
automated build system seems to use that as the working directory, and
so fails at the last step (`chmod +x index.js && ...`) with "no such
file or directory" errors.

(I can get it work when building locally by using the -f flag to
`docker build`, but I dont' see how to set this in the Docker Hub
automated build config).